+++ src/pretalx/agenda/management/commands/export_schedule_html.py
@@ -78,11 +78,13 @@ def event_exporter_urls(event):
 def schedule_version_urls(event):
     for schedule in event.schedules.filter(version__isnull=False):
         yield schedule.urls.public
+        yield schedule.urls.widget_data
 
 
 def event_urls(event):
     yield event.urls.base
     yield event.urls.schedule
+    yield event.urls.widget_data
     yield from schedule_version_urls(event)
     yield event.urls.featured
     yield event.urls.talks