+++ src/pretalx/agenda/urls.py
@@ -23,11 +23,8 @@ def get_schedule_urls(regex_prefix, name_prefix=""):
             (".xcal", schedule.ExporterView.as_view(), "export.schedule.xcal"),
             (".json", schedule.ExporterView.as_view(), "export.schedule.json"),
             (".ics", schedule.ExporterView.as_view(), "export.schedule.ics"),
-            (
-                "/export/<name>",
-                schedule.ExporterView.as_view(),
-                "export",
-            ),
+            ("/widget/v2.json", widget.widget_data_v2, "widget.data.2"),
+            ("/export/<name>", schedule.ExporterView.as_view(), "export"),
         ]
     ]
 
@@ -63,11 +60,6 @@ urlpatterns = [
                     "schedule/widget/v1.json",
                     widget.WidgetData.as_view(),
                     name="widget.data",
-                ),
-                path(
-                    "schedule/widget/v2.json",
-                    widget.widget_data_v2,
-                    name="widget.data.2",
                 ),
                 *get_schedule_urls("schedule"),
                 *get_schedule_urls("schedule/v/<version>", "versioned-"),