+++ src/pretalx/schedule/models/schedule.py
@@ -55,6 +55,7 @@ class Schedule(LogMixin, models.Model):
 
     class urls(EventUrls):
         public = "{self.event.urls.schedule}v/{self.url_version}/"
+        widget_data = "{public}widget/v2.json"
 
     @transaction.atomic
     def freeze(